refactor : RunningService의 중복 코드를 줄이기 위한 BaseRunningService 추상 클래스 적용 #PAR-335
