+++ app/src/main/AndroidManifest.xml
@@ -43,7 +43,9 @@
             android:theme="@style/Theme.PartyRunApplication"
             android:launchMode="singleTop">
         </activity>
-
+        <service
+            android:name=".feature.running.service.BaseRunningService"
+            android:foregroundServiceType="location" />
         <service
             android:name=".feature.running.service.BattleRunningService"
             android:foregroundServiceType="location" />
